Merging
Former-commit-id: 3bf8d5995a64d59f9ba9ef33431600a52ea09e5f
Former-commit-id: a3442abd5ea87d38c98137d78ba00ddf08fc9e71

+++ src/rendering/selectionRenderPass.cpp
@@ -4,7 +4,7 @@
 
 namespace phi
 {
-    selectionRenderPass::selectionRenderPass(const gBufferRenderPass* const gBufferPass, const gl* const gl, size_t w, size_t h) :
+    selectionRenderPass::selectionRenderPass(const gBufferRenderPass* const gBufferPass, const gl* const gl, float w, float h) :
         _gl(gl),
         _quad(nullptr),
         _quadVbo(nullptr),
